Fix depedency issue
