+++ src/domain/submission/types.ts
@@ -50,3 +50,13 @@ export type FileDetails = {
     manuscriptFile?: File | null; // responsibility of the Files Service
     supportingFiles?: Array<File>; // responsibility of the Files Service
 };
+
+
+export enum SubmissionStatus {
+    INITIAL = 'INITIAL',
+    MECA_EXPORT_PENDING = 'MECA_EXPORT_PENDING',
+    MECA_EXPORT_FAILED = 'MECA_EXPORT_FAILED',
+    MECA_EXPORT_SUCCEEDED = 'MECA_EXPORT_SUCCEEDED',
+    MECA_IMPORT_FAILED = 'MECA_IMPORT_FAILED',
+    MECA_IMPORT_SUCCEEDED = 'MECA_IMPORT_SUCCEEDED',
+}
